python-scripts/requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
pandas
pandas-gbq
requests
schedule
python-dotenv
streamlit
db-dtypes
plotly
numpy
pytz
